msm: kgsl: Don't cast pwr->max_pwrlevel

+++ drivers/gpu/msm/kgsl_pwrscale.c
@@ -330,11 +330,7 @@ int kgsl_devfreq_target(struct device *dev, unsigned long *freq, u32 flags)
 	/* If the governor recommends a new frequency, update it here */
 	if (*freq != cur_freq) {
 		level = pwr->max_pwrlevel;
-		/*
-		 * To avoid infinite loop issue type cast max_pwrlevel to
-		 * signed integer type
-		 */
-		for (i = pwr->min_pwrlevel; i >= (int)pwr->max_pwrlevel; i--)
+		for (i = pwr->min_pwrlevel; i >= pwr->max_pwrlevel; i--)
 			if (*freq <= pwr->pwrlevels[i].gpu_freq) {
 				if (pwr->thermal_cycle == CYCLE_ACTIVE)
 					level = _thermal_adjust(pwr, i);
